Entity: Use @ElementCollection instead of @Basic
The Basic value collection, was not read correctly by whatever JPA
casted the value to. This commit fixes this by changing from the @Basic
annotation to an @ElementCollection. @ElementCollection only takes the
abstract List type as a single value, so we need to redefine all values
from ArrayList to List. The initialization in the entity ensures that we
still get the functionality of the ArrayList.