FS20G06 Add Game service with createGame endpoint
This commit implements the start game business logic. It creates a new
game entity, fills it with the relavant information and moves users to
the game. It also populates a word list for the game based on the
english word list text file.

Also implements unit tests for these changes.